Handle potential @db_fields['CallableQueries']
This code change enables handling of  `@db_fields['CallableQueries']`.

The `config.db_fields` hash is set in `config/initializers/translation.rb` of the codebase using this gem. Within `config.db_fields`, if the 'CallableQueries' key is added, then the value must be an array of callable ActiveRecord query procs. Those query procs are then executed here and their returned results are added to the entries array.